tests/wpt/web-platform-tests/trusted-types/HTMLObjectElement.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  //helper function for the tests
  function testData(str, url) {
    var objectElement = document.createElement('object');
    objectElement.data = url;
    objectElement.codeBase = url;
    assert_equals(objectElement.data, str);
    assert_equals(objectElement.codeBase,str);
  }

  test(t => {
    testData(URLS.safe, TrustedURL.create(URLS.safe));
  }, "Basic processing: safe URL, safe construction.");

  test(t => {
    testData(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
  }, "Basic processing: safe URL, unsafe construction.");
</script>
